Separate API responsibilities a bit more cleanly
As part of #100
